ai_pipelines/requirements.txt
# Google ADK Agents Requirements for MediSense

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Vector Database
chromadb>=0.4.0

# Image Processing
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0

# Logging and Utilities
python-dotenv>=1.0.0
